Split out kdf key to separate file
