+++ crates/bitwarden-crypto/src/keys/mod.rs
@@ -1,10 +1,7 @@
 mod key_encryptable;
 pub use key_encryptable::{CryptoKey, KeyContainer, KeyDecryptable, KeyEncryptable};
 mod master_key;
-pub use master_key::{
-    default_argon2_iterations, default_argon2_memory, default_argon2_parallelism,
-    default_pbkdf2_iterations, HashPurpose, Kdf, MasterKey,
-};
+pub use master_key::{HashPurpose, MasterKey};
 mod shareable_key;
 pub use shareable_key::derive_shareable_key;
 mod symmetric_crypto_key;
@@ -21,4 +18,9 @@ mod device_key;
 pub use device_key::{DeviceKey, TrustDeviceResponse};
 mod pin_key;
 pub use pin_key::PinKey;
+mod kdf;
+pub use kdf::{
+    default_argon2_iterations, default_argon2_memory, default_argon2_parallelism,
+    default_pbkdf2_iterations, Kdf,
+};
 mod utils;
