+++ crates/bitwarden-crypto/src/keys/pin_key.rs
@@ -1,10 +1,10 @@
-use super::master_key::{decrypt_user_key, encrypt_user_key, KdfDerivedKeyMaterial};
+use super::{
+    kdf::{Kdf, KdfDerivedKeyMaterial},
+    master_key::{decrypt_user_key, encrypt_user_key},
+};
 use crate::{
-    keys::{
-        key_encryptable::CryptoKey,
-        utils::{derive_kdf_key, stretch_kdf_key},
-    },
-    EncString, Kdf, KeyEncryptable, Result, SymmetricCryptoKey,
+    keys::{key_encryptable::CryptoKey, utils::stretch_key},
+    EncString, KeyEncryptable, Result, SymmetricCryptoKey,
 };
 
 /// Pin Key.
@@ -15,17 +15,17 @@ pub struct PinKey(KdfDerivedKeyMaterial);
 impl PinKey {
     /// Derives a users pin key from their password, email and KDF.
     pub fn derive(password: &[u8], email: &[u8], kdf: &Kdf) -> Result<Self> {
-        derive_kdf_key(password, email, kdf).map(Self)
+        KdfDerivedKeyMaterial::derive_kdf_key(password, email, kdf).map(Self)
     }
 
     /// Encrypt the users user key
     pub fn encrypt_user_key(&self, user_key: &SymmetricCryptoKey) -> Result<EncString> {
-        encrypt_user_key(&self.0, user_key)
+        encrypt_user_key(&self.0.key_material, user_key)
     }
 
     /// Decrypt the users user key
     pub fn decrypt_user_key(&self, user_key: EncString) -> Result<SymmetricCryptoKey> {
-        decrypt_user_key(&self.0, user_key)
+        decrypt_user_key(&self.0.key_material, user_key)
     }
 }
 
@@ -33,7 +33,7 @@ impl CryptoKey for PinKey {}
 
 impl KeyEncryptable<PinKey, EncString> for &[u8] {
     fn encrypt_with_key(self, key: &PinKey) -> Result<EncString> {
-        let stretched_key = SymmetricCryptoKey::Aes256CbcHmacKey(stretch_kdf_key(&key.0)?);
+        let stretched_key = SymmetricCryptoKey::Aes256CbcHmacKey(stretch_key(&key.0.key_material)?);
         self.encrypt_with_key(&stretched_key)
     }
 }
